finished fixing test

+++ part4/blog-list/controllers/blogs.js
@@ -1,7 +1,5 @@
 const blogsRouter = require("express").Router()
 const Blog = require("../models/blog")
-const User = require("../models/user")
-const jwt = require("jsonwebtoken")
 
 blogsRouter.get("/", async (req, res) => {
   const blogs = await Blog.find({}).populate("user", {
@@ -9,17 +7,12 @@ blogsRouter.get("/", async (req, res) => {
     name: 1,
     id: 1,
   })
-  res.json(blogs)
+  res.json(blogs.map((blog) => blog.toJSON()))
 })
 
 blogsRouter.post("/", async (req, res, next) => {
   const body = req.body
-  const decodedToken = jwt.verify(req.token, process.env.SECRET)
-
-  if (!decodedToken.id) {
-    return res.status(401).json({ error: "token invalid" })
-  }
-  const user = await User.findById(decodedToken.id)
+  const user = req.user
 
   if (!body.title || !body.url) {
     return res.status(400).json({ error: "title and url requried" })
@@ -40,6 +33,19 @@ blogsRouter.post("/", async (req, res, next) => {
 })
 
 blogsRouter.delete("/:id", async (req, res) => {
+  const user = req.user
+  const blog = await Blog.findById(req.params.id)
+
+  if (!blog) {
+    return res.status(404).json({ error: "blog not found" })
+  }
+
+  if (blog.user._id.toString() !== user._id.toString()) {
+    return res
+      .status(401)
+      .json({ error: "unauthorized only creator can delete this blog" })
+  }
+
   await Blog.findByIdAndDelete(req.params.id)
   res.status(204).end()
 })
